+++ moderatorFrontend/src/jsMain/kotlin/views/common/genericErrorView/GenericErrorViewConfig.kt
@@ -1,0 +1,6 @@
+package views.common.genericErrorView
+
+class GenericErrorViewConfig(
+  val title: String = "",
+  val subtitle: String = "",
+)
