+++ moderatorFrontend/src/jsMain/kotlin/views/guestCheckIn/addGuestCheckIn/AddGuestCheckInConfig.kt
@@ -1,0 +1,9 @@
+package views.guestCheckIn.addGuestCheckIn
+
+import react.MutableRefObject
+import webcore.MbDialogRef
+
+class AddGuestCheckInConfig(
+  val dialogRef: MutableRefObject<MbDialogRef>,
+  val onGuestCheckedIn: () -> Unit,
+)
