+++ moderatorFrontend/src/jsMain/kotlin/views/guestCheckIn/addGuestCheckIn/AddGuestCheckInController.kt
@@ -1,4 +1,4 @@
-package views.guestCheckIn
+package views.guestCheckIn.addGuestCheckIn
 
 import app.appContextToInject
 import app.baseUrl
