+++ moderatorFrontend/src/jsMain/kotlin/views/guestCheckIn/guestCheckInRow/GuestCheckInRowConfig.kt
@@ -1,0 +1,8 @@
+package views.guestCheckIn.guestCheckInRow
+
+import com.studo.campusqr.common.payloads.ActiveCheckIn
+
+class GuestCheckInRowConfig(
+  val activeCheckIn: ActiveCheckIn,
+  val onCheckedOut: () -> Unit,
+)
