+++ moderatorFrontend/src/jsMain/kotlin/views/login/LoginViewConfig.kt
@@ -1,0 +1,11 @@
+package views.login
+
+class LoginViewConfig(
+  val loginMode: LoginMode,
+) {
+  companion object {
+    enum class LoginMode {
+      EMAIL
+    }
+  }
+}
