+++ moderatorFrontend/src/jsMain/kotlin/views/login/loginNavigationButtonsViewFc/LoginNavigationButtonsViewConfig.kt
@@ -1,0 +1,9 @@
+package views.login.loginNavigationButtonsViewFc
+
+class LoginNavigationButtonsViewConfig(
+  var networkRequestInProgress: Boolean,
+  var backEnabled: Boolean,
+  var nextButtonText: String,
+  var nextButtonDisabled: Boolean,
+  var onNextAction: () -> Unit
+)
