+++ moderatorFrontend/src/jsMain/kotlin/views/users/listUsersToolbarView/ListUsersToolbarViewConfig.kt
@@ -1,0 +1,9 @@
+package views.users.listUsersToolbarView
+
+import react.MutableRefObject
+import webcore.MbDialogRef
+
+class ListUsersToolbarViewConfig(
+  val dialogRef: MutableRefObject<MbDialogRef>,
+  val handleCreateOrAddUserResponse: (String?) -> Unit,
+)
